Extract Footer.js to folder with TwitterWidget
  I pulled this out for the UI mockup which shows these side by side.
  If this is interpreted incorrect we can separate them.
