+++ src/components/footer/Footer.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import {StyleSheet, css} from 'aphrodite';
+import TwitterWidget from './TwitterWidget';
 
 const s = StyleSheet.create({
   footer: {
@@ -24,15 +25,13 @@ const s = StyleSheet.create({
 });
 
 
-const Footer = () => {
-    return (
-        <div className={css(s.footer)}>
-            <a className={css(s.a)} href="https://docs.google.com/document/d/1uQda1goIN5m1Rb5anVNlWgU7Ra-uBEZBIa1GaHqfRH4/edit?usp=sharing" target="_blank">
-              <div className={css(s.about)}>About</div>
-              <img className={css(s.logo)} alt='matchus logo' src='./logo.png' />
-            </a>
-        </div>
-    );
-};
-
-export default Footer;
+export default () => 
+  <div>
+    <TwitterWidget />
+    <div className={css(s.footer)}>
+        <a className={css(s.a)} href="https://docs.google.com/document/d/1uQda1goIN5m1Rb5anVNlWgU7Ra-uBEZBIa1GaHqfRH4/edit?usp=sharing" target="_blank">
+          <div className={css(s.about)}>About</div>
+          <img className={css(s.logo)} alt='matchus logo' src='./logo.png' />
+        </a>
+    </div>
+  </div>
